src/CMS/src/pyproject.toml
[project]
name = "sms-cms"
version = "2.0.0"
description = "Central Management System for SMS Arduino node"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-socketio>=5.3.6",
    "pyserial>=3.5",
    "eventlet>=0.35.1",
    # sqlite3 is part of the Python standard library — no extra package needed
]

[project.scripts]
cms = "app:main"
